// packages/adricore/src/metadata/enums.ts
export enum ObjectFieldDataTypes {
  Id = "Id",
  String = "String",
  Boolean = "Boolean",
  Integer = "Integer",
  Decimal = "Decimal",
  Date = "Date",
  Time = "Time",
  Datetime = "Datetime",
  TextArea = "TextArea",
  Reference = "Reference",
  Email = "Email",
  Phone = "Phone",
  Url = "Url",
  Picklist = "Picklist",
  MultiPicklist = "MultiPicklist",
  Code = "Code"
}

export enum ObjectRelationshipReferenceTypes {
  MasterDetail = "MasterDetail",
  Lookup = "Lookup"
}